requirements.txt
Flask==2.3.2
gTTS==2.3.2
Pillow==10.2.0
